Assistente_Virtual_ALICE/requirements.txt
customtkinter
Pillow
speechrecognition
pyttsx3
pyparsing
tkinterdnd2
langchain
nltk
